feat: reorganize Features section and implement new FeatureCard component with data-driven approach
